feat: big-map: refactor map-status-bar in separate file
